requirements-dev.txt
# TemplateHub - Development Requirements
# Includes all minimal requirements plus development tools

# Include minimal requirements
-r requirements-minimal.txt

# Development and Debugging Tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Testing Framework
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2

# Code Quality Tools
flake8==6.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Documentation
Sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Database Tools (for development)
django-seed==0.3.1

# API Development
djangorestframework==3.15.2
drf-spectacular==0.27.0

# Environment Management
python-decouple==3.8

# Installation Instructions for Development:
# 1. Create virtual environment: python -m venv venv
# 2. Activate virtual environment: 
#    - Windows: venv\Scripts\activate
#    - macOS/Linux: source venv/bin/activate
# 3. Install dev requirements: pip install -r requirements-dev.txt
# 4. Run migrations: python manage.py migrate
# 5. Create superuser: python manage.py createsuperuser
# 6. Run tests: pytest
# 7. Run server: python manage.py runserver
